+++ py_trees/behaviours.py
@@ -14,10 +14,12 @@ A library of fundamental behaviours for use.
 # Imports
 ##############################################################################
 
+import functools
 import operator
 import typing
 
 from . import behaviour
+from . import blackboard
 from . import common
 from . import meta
 
@@ -141,6 +143,53 @@ class SuccessEveryN(behaviour.Behaviour):
         else:
             self.feedback_message = "not yet"
             return common.Status.FAILURE
+
+
+class TickCounter(behaviour.Behaviour):
+    """
+    A useful utility behaviour for demos and tests. Simply
+    ticks with :data:`~py_trees.common.Status.RUNNING` for
+    the specified number of ticks before returning the
+    requested completion status (:data:`~py_trees.common.Status.SUCCESS`
+    or :data:`~py_trees.common.Status.FAILURE`).
+
+    This behaviour will reset the tick counter when initialising.
+
+    Args:
+        name: name of the behaviour
+        duration: number of ticks to run
+        completion_status: status to switch to once the counter has expired
+    """
+    def __init__(
+        self,
+        duration: int,
+        name=common.Name.AUTO_GENERATED,
+        completion_status: common.Status=common.Status.SUCCESS
+    ):
+        super().__init__(name=name)
+        self.completion_status = completion_status
+        self.duration = duration
+        self.counter = 0
+
+    def initialise(self):
+        """
+        Reset the tick counter.
+        """
+        self.counter = 0
+
+    def update(self):
+        """
+        Increment the tick counter and return the appropriate status for this behaviour
+        based on the tick count.
+
+        Returns
+            :data:`~py_trees.common.Status.RUNNING` while not expired, the given completion status otherwise
+        """
+        self.counter += 1
+        if self.counter <= self.duration:
+            return common.Status.RUNNING
+        else:
+            return self.completion_status
 
 
 class Count(behaviour.Behaviour):
@@ -404,8 +453,6 @@ class CheckBlackboardVariableValue(behaviour.Behaviour):
 
     .. tip::
         The python `operator module`_ includes many useful comparison operations.
-
-    .. _`operator module`: https://docs.python.org/2/library/operator.html
     """
     def __init__(
             self,
@@ -504,3 +551,83 @@ class WaitForBlackboardVariableValue(CheckBlackboardVariableValue):
             return common.Status.RUNNING
         else:
             return new_status
+
+
+class CheckBlackboardVariableValues(behaviour.Behaviour):
+    """
+    Apply a logical operation across a set of blackboard variable checks.
+    This is non-blocking, so will always tick with status
+    :data:`~py_trees.common.Status.FAILURE` or
+    :data:`~py_trees.common.Status.SUCCESS`.
+
+    Args:
+        checks: a list of comparison checks to apply to blackboard variables
+        logical_operator: a logical check to apply across the results of the blackboard variable checks
+        name: name of the behaviour
+        namespace: optionally store results of the checks (boolean) under this namespace
+
+    .. tip::
+        The python `operator module`_ includes many useful logical operators, e.g. operator.xor.
+
+    Raises:
+        ValueError if less than two variable checks are specified (insufficient for logical operations)
+    """
+    def __init__(
+        self,
+        checks: typing.List[common.ComparisonExpression],
+        operator: typing.Callable[[bool, bool], bool],
+        name: str=common.Name.AUTO_GENERATED,
+        namespace: str=None,
+    ):
+        super().__init__(name=name)
+        self.checks = checks
+        self.operator = operator
+        self.blackboard = self.attach_blackboard_client()
+        if len(checks) < 2:
+            raise ValueError("Must be at least two variables to operate on [only {} provided]".format(len(checks)))
+        for check in self.checks:
+            self.blackboard.register_key(
+                key=blackboard.Blackboard.key(check.variable),
+                access=common.Access.READ
+            )
+        if namespace is not None:
+            self.blackboard_results = self.attach_blackboard_client(namespace=namespace)
+            for counter in range(1, len(self.checks) + 1):
+                self.blackboard_results.register_key(
+                    key=str(counter),
+                    access=common.Access.WRITE
+                )
+        else:
+            self.blackboard_results = None
+
+    def update(self) -> common.Status:
+        """
+        Applies comparison checks on each variable and a logical check across the
+        complete set of variables.
+
+        Returns:
+             :data:`~py_trees.common.Status.FAILURE` if key retrieval or logical checks failed, :data:`~py_trees.common.Status.SUCCESS` otherwise.
+        """
+        self.logger.debug("%s.update()" % self.__class__.__name__)
+        results = []
+        for check in self.checks:
+            try:
+                value = self.blackboard.get(check.variable)
+            except KeyError:
+                self.feedback_message = "variable '{}' does not yet exist on the blackboard".format(self.variable_name)
+                return common.Status.FAILURE
+            results.append(check.operator(value, check.value))
+        if self.blackboard_results is not None:
+            for counter in range(1, len(results) + 1):
+                self.blackboard_results.set(str(counter), results[counter - 1])
+        logical_result = functools.reduce(self.operator, results)
+        if logical_result:
+            self.feedback_message = "[{}]".format(
+                "|".join(["T" if result else "F" for result in results])
+            )
+            return common.Status.SUCCESS
+        else:
+            self.feedback_message = "[{}]".format(
+                "|".join(["T" if result else "F" for result in results])
+            )
+            return common.Status.FAILURE